OrderService/pyproject.toml
[tool.poetry]
name = "orderservice"
version = "0.1.0"
description = ""
authors = ["sumets <svetasumets@gmail.com>"]
readme = "README.md"

[tool.poetry.dev-dependencies]
ruff = "^0.3.0"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
beanie = "^1.25.0"
pydantic-settings = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pymongo = "^4.6.2"
httpx = "^0.27.0"

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [".venv", "migrations/"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
select = [
  "E", # Errors
  "F", # Pyflakes
  "I", # isort
  "T20", # flake8-print
]

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = true
case-sensitive = false
detect-same-package = true
order-by-type = false
force-sort-within-sections = true
known-local-folder = ["tests", "app"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
